+++ frb_codegen/src/library/codegen/generator/api_dart/mod.rs
@@ -71,6 +71,20 @@ mod tests {
 
     #[test]
     #[serial]
+    fn test_simple_dart_collection_deep_equality() -> anyhow::Result<()> {
+        body_with_config(
+            "library/codegen/generator/api_dart/mod/simple",
+            HashMap::from([
+                ("api.dart", "expect_output.dart"),
+                ("dep.dart", "expect_output2_deep_collection_equality.dart"),
+                ("frb_generated.dart", "expect_output3.dart"),
+            ]),
+            Some(true),
+        )
+    }
+
+    #[test]
+    #[serial]
     fn test_functions() -> anyhow::Result<()> {
         body(
             "library/codegen/generator/api_dart/mod/functions",
@@ -82,11 +96,20 @@ mod tests {
     }
 
     fn body(fixture_name: &str, expect_outputs: HashMap<&str, &str>) -> anyhow::Result<()> {
+        body_with_config(fixture_name, expect_outputs, None)
+    }
+
+    fn body_with_config(
+        fixture_name: &str,
+        expect_outputs: HashMap<&str, &str>,
+        dart_collection_deep_equality: Option<bool>,
+    ) -> anyhow::Result<()> {
         configure_opinionated_test_logging();
         let test_fixture_dir = get_test_fixture_dir(fixture_name);
         env::set_current_dir(&test_fixture_dir)?;
 
-        let config = Config::from_files_auto()?;
+        let mut config = Config::from_files_auto()?;
+        config.dart_collection_deep_equality = dart_collection_deep_equality;
         let internal_config = InternalConfig::parse(&config, &MetaConfig { watch: false })?;
         let mir_pack = crate::codegen::parser::parse(
             &internal_config.parser,