+++ frb_codegen/test_fixtures/library/codegen/generator/api_dart/mod/simple/src/dep.rs
@@ -1,3 +1,4 @@
-pub struct Simple{
-  val:u32
+pub struct Simple {
+    val: u32,
+    values: Vec<String>,
 }